move around package structure
